Module analysis::mir_utils

Various helper functions for working with mir types. copied from prusti-interface/utils

Structs

Place

A wrapper for mir::Place that implements Ord.

Traits

PlaceImpl

A trait enabling Place and mir::Place to be treated in the same way

Functions

expand_struct_place

Expands a place x.f.g of type struct into a vector of places for each of the struct’s fields {x.f.g.f, x.f.g.g, x.f.g.h}. If without_field is not None, then omits that field from the final vector.

get_blocked_place

is_copy

location_to_stmt_str

Convert a location to a string representing the statement or terminator at that location

remove_place_from_set

Remove all extensions of a place from a set, unpacking prefixes as much as needed.