Function analysis::mir_utils::expand_struct_place

pub fn expand_struct_place<'tcx, P: PlaceImpl<'tcx> + Copy>(
    place: P,
    mir: &Body<'tcx>,
    tcx: TyCtxt<'tcx>,
    without_field: Option<usize>
) -> Vec<P>

Expands a place x.f.g of type struct into a vector of places for each of the struct’s fields {x.f.g.f, x.f.g.g, x.f.g.h}. If without_field is not None, then omits that field from the final vector.