Struct analysis::PointwiseState

pub struct PointwiseState<'mir, 'tcx: 'mir, S: Serialize> { /* private fields */ }

Records the state of the analysis at every program point and CFG edge of mir.

Implementations

impl<'mir, 'tcx: 'mir> PointwiseState<'mir, 'tcx, DefinitelyAccessibleState<'tcx>>

pub fn generate_test_program( &self, tcx: TyCtxt<'tcx>, source_map: &SourceMap ) -> String

Make a best-effort at injecting statements to check the analysis state.

impl<'mir, 'tcx: 'mir, S: Serialize> PointwiseState<'mir, 'tcx, S>

pub fn new(mir: &'mir Body<'tcx>) -> Self

pub fn lookup_before(&self, location: Location) -> Option<&S>

Look up the state before the location. The location can point to a statement or terminator.

pub fn lookup_mut_before(&mut self, location: Location) -> Option<&mut S>

Look up the mutable state before the location. The location can point to a statement or terminator.

pub fn lookup_after(&self, location: Location) -> Option<&S>

Look up the state after the location. The location should point to a statement, not a terminator.

pub fn lookup_after_block( &self, block: BasicBlock ) -> Option<&FxHashMap<BasicBlock, S>>

Look up the state on the outgoing CFG edges of block. The return value maps all successor blocks to the state on the CFG edge from block to that block.

impl<'mir, 'tcx: 'mir, S: Serialize + Default> PointwiseState<'mir, 'tcx, S>

pub fn default(mir: &'mir Body<'tcx>) -> Self