Trait analysis::Analysis

pub trait Analysis<'mir, 'tcx: 'mir> {
    type State: AbstractState;
    fn def_id(&self) -> DefId;

    fn body(&self) -> &'mir Body<'tcx>;

    fn new_bottom(&self) -> Self::State;

    fn new_initial(&self) -> Self::State;

    fn need_to_widen(counter: u32) -> bool;

    fn apply_statement_effect(
        &self, 
        state: &mut Self::State, 
        location: Location
    ) -> Result<(), AnalysisError>;

    fn apply_terminator_effect(
        &self, 
        state: &Self::State, 
        location: Location
    ) -> Result<Vec<(BasicBlock, Self::State)>, AnalysisError>;

    fn run_fwd_analysis(
        &self
    ) -> Result<PointwiseState<'mir, 'tcx, Self::State>, AnalysisError> { ... }
}

Trait to be used to define the static analysis of a MIR body.

Associated Types

type State: AbstractState

Required methods

fn def_id(&self) -> DefId

Return the DefId of the MIR body to be analyzed.

fn body(&self) -> &'mir Body<'tcx>

Return the MIR body to be analyzed.

fn new_bottom(&self) -> Self::State

Creates a new abstract state which corresponds to the bottom element in the lattice

fn new_initial(&self) -> Self::State

Creates the abstract state at the beginning of the mir body. In particular this should take the arguments into account.

fn need_to_widen(counter: u32) -> bool

Determines if the number of times a block was traversed by the analyzer given in counter is large enough to widen the state

fn apply_statement_effect(
&self,
state: &mut Self::State,
location: Location
) -> Result<(), AnalysisError>

Modify a state according to the statement at location.

The statement can be extracted using self.mir[location.block].statements[location.statement_index].

fn apply_terminator_effect(
&self,
state: &Self::State,
location: Location
) -> Result<Vec<(BasicBlock, Self::State)>, AnalysisError>

Compute the states after a terminator at location.

The statement can be extracted using self.mir[location.block].terminator().

Provided methods

fn run_fwd_analysis(
&self
) -> Result<PointwiseState<'mir, 'tcx, Self::State>, AnalysisError>

Produces an abstract state for every program point in mir by iterating over all statements in program order until a fixed point is reached (i.e. by abstract interpretation).